refactor clip command package
